requirements.txt
asgiref==3.8.1
black==25.1.0
cfgv==3.4.0
click==8.1.8
colorama==0.4.6
distlib==0.3.9
Django==5.2.1
django-filter==25.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-yasg==1.21.10
filelock==3.18.0
identify==2.6.10
inflection==0.5.1
mypy_extensions==1.1.0
nodeenv==1.9.1
packaging==25.0
pathspec==0.12.1
platformdirs==4.3.8
pre_commit==4.2.0
PyJWT==2.9.0
python-decouple==3.8
pytz==2025.2
PyYAML==6.0.2
sqlparse==0.5.3
tomli==2.2.1
typing_extensions==4.13.2
tzdata==2025.2
uritemplate==4.1.1
virtualenv==20.31.1
